+++ NEWS.md
@@ -1,6 +1,9 @@
 # CFAEpiNow2Pipeline (development version)
 
 * Populated the default values of the metadata to be saved.
+* Creating a Config class to make syncing configuration differences easier.
+* Add a JSON reader for the Config class.
+* Use the Config class throughout the pipeline.
 * Adding a script to setup the Azure Batch Pool to link the container.
 * Adding new action to post a comment on PRs with a link to the rendered pkgdown site.
 * Add inner pipeline responsible for running the model fitting process